pyproject.toml
[project]
name = "google-adk-research-agent"
version = "0.1.0"
description = "Multi-agent research pipeline using Google ADK with LangSmith tracing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.28.0",
    "langsmith[google-adk]>=0.7.23",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
